Update README.md
It’s the show we give you some straight-up gameplay and our first impressions of the latest games releasing and as usual. It’s me Jilon Tanvir, we’re talking about Hitman 3 Today I’ll come right out and say it. And then three here is the best of this new trilogy of Hitman games.
https://gsmteng.com/hitman-3/

The world of assassination trilogy as it’s called, and just a heads up to this video doesn’t have spoilers but it does have a few kills in location so be aware if you want to go in blind. Also, this was captured on PC, yes this is a lot of video of just a man walking around in rooms, but that’s how Hitman goes so buckle up just for some context. I want to be clear, I’m coming at this as a very specific type of fan.

Description
Hitman 3, the world’s greatest assassin accepts his final contract. In this dramatic finale of the stealth-action series from IO Interactive. In this third major installment, players will venture into new locations and take out there. Targets with surgical precision, etc release this 20th of January 2021